ClaudeSpeak/claude-voice-connector-http/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
starlette>=0.37
pydantic>=2.0
piper-tts>=1.2
onnxruntime>=1.17
pyyaml>=6.0
numpy>=1.24,<2
mcp>=1.9
sounddevice>=0.4
# coqui-tts provides XTTS-v2.  Large install (~2.5 GB incl. torch), but gives
# us the higher-quality voice engine the user addendum requested.
# Pin transformers<5 because coqui-tts 0.27 calls `transformers.pytorch_utils
# .isin_mps_friendly`, which was removed in transformers 5.x.
coqui-tts>=0.24
transformers>=4.39,<5
